UI fixes added
